Fix setup on boxes app store based WSL install (#7)
## Summary of the Pull Request

The installer will fail on windows boxes where WSL has been installed
via the app store/OOB and not as a windows features.
This is because the Linux kernel is placed in a different location for
ap store installations and the `wsl_update_x64.msi` installer will not
recognize that WSL is installed.

This update the configuration manifest to use the "new" way to
install/update WSL (Windows 10 version 2004 Build >= 19041 or Windows
11) instead of installing it manually via windows features + the msi
(`wsl_update_x64.msi`)
> Source:
https://learn.microsoft.com/en-us/windows/wsl/install#install-wsl-command

## Detailed Description of the Pull Request / Additional comments

Unfortunately we need to make this optional as it's not trivial to
handle all possible use cases correctly.

Note that this will only work if the user either
- Did not install WSL yet
- Already installed WSL from the app store
- Already installed WSL using `wsl.exe --install ...`

It will fail if the user already installed WSL using the "inbox" way
(enabled through windows features).
In this case the wsl.exe CLI will not have the --no-distribution and
--web-download parameters.

If the user experience issues, it can uncheck the WSL update steps and
then can update/install WSL himself without being blocked.

Co-authored-by: Luca Piombino <luca.piombino@codecoil.com>