Bump docker binaries versions (docker: `28.2.2`, compose: `2.36.2`, buildx: `0.24.0`) (#10)
## Summary of the Pull Request
Update docker (and related tooling) to the latest version.

## Detailed Description of the Pull Request / Additional comments

The core motivation is to have a version of the docker client that does
not attempt to mutate/rewrite the config
(https://github.com/docker/cli/pull/5553).

This is in preparation of a future enhancement to share the same client
configuration of the windows host and to provide a seamless experience.
This could include also the auth via credential helpers running on the
host (e.g. a basic scenario is to at least use the windows credential
manager to avoid storing static credentials in plain in the config
file).

I did not push anything yet because i am waiting to see if
https://github.com/docker/cli/pull/5948
get merged as it would probably provide a better and simpler out of the 
box experience.

Co-authored-by: Luca Piombino <luca.piombino@codecoil.com>